+++ packages/brick_build/lib/src/builders/sqlite_base_builder.dart
@@ -1,6 +1,5 @@
 import 'package:build/build.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:brick_build/src/annotation_super_generator.dart';
 import 'package:brick_build/src/builders/base.dart';
 import 'package:brick_build/src/offline_first/sqlite_serdes.dart';
 import 'package:brick_build/src/sqlite_schema/sqlite_schema_generator.dart';
@@ -10,10 +9,10 @@ export 'package:brick_build/src/annotation_super_generator.dart';
 
 const _schemaGenerator = SqliteSchemaGenerator();
 
-abstract class SqliteBaseBuilder extends BaseBuilder {
+abstract class SqliteBaseBuilder<_ClassAnnotation> extends BaseBuilder<_ClassAnnotation> {
   SqliteSchemaGenerator get schemaGenerator => _schemaGenerator;
 
-  SqliteBaseBuilder(AnnotationSuperGenerator generator) : super(generator);
+  SqliteBaseBuilder();
 
   Future<List<SqliteFields>> sqliteFieldsFromBuildStep(BuildStep buildStep) async {
     final annotatedElements = await getAnnotatedElements(buildStep);