+++ packages/brick_build/lib/src/sqlite_serdes/sqlite_deserialize.dart
@@ -1,5 +1,6 @@
 import 'package:analyzer/dart/element/element.dart' show ClassElement;
 import 'package:brick_build/src/offline_first/offline_first_serdes_generator.dart';
+import 'package:brick_build/src/serdes_generator.dart';
 import 'package:brick_build/src/sqlite_serdes/sqlite_fields.dart';
 import 'package:source_gen/source_gen.dart' show InvalidGenerationSourceError;
 import 'package:brick_sqlite_abstract/db.dart' show InsertTable;
@@ -25,9 +26,9 @@ class SqliteDeserialize extends OfflineFirstSerdesGenerator<Sqlite> {
       "..${InsertTable.PRIMARY_KEY_FIELD} = data['${InsertTable.PRIMARY_KEY_COLUMN}'] as int;";
 
   @override
-  String coderForField({checker, offlineFirstAnnotation, wrappedInFuture, field, fieldAnnotation}) {
-    final name = serializedFieldName(checker, fieldAnnotation.name);
-    final defaultValue = defaultValueSuffix(fieldAnnotation);
+  String coderForField(field, checker, {offlineFirstAnnotation, wrappedInFuture, fieldAnnotation}) {
+    final fieldValue = serdesValueForField(field, fieldAnnotation.name, checker: checker);
+    final defaultValue = SerdesGenerator.defaultValueSuffix(fieldAnnotation);
     if (field.name == InsertTable.PRIMARY_KEY_FIELD) {
       throw InvalidGenerationSourceError(
         'Field `${InsertTable.PRIMARY_KEY_FIELD}` conflicts with reserved `SqliteModel` getter.',
@@ -36,27 +37,23 @@ class SqliteDeserialize extends OfflineFirstSerdesGenerator<Sqlite> {
       );
     }
 
-    if (fieldAnnotation?.fromGenerator != null) {
-      return digestPlaceholders(fieldAnnotation.fromGenerator, name, field.name);
-    }
-
     // DateTime
     if (checker.isDateTime) {
-      return "data['$name'] == null ? null : DateTime.tryParse(data['$name']$defaultValue as String)";
+      return "$fieldValue == null ? null : DateTime.tryParse($fieldValue$defaultValue as String)";
 
       // bool
     } else if (checker.isBool) {
-      return "data['$name'] == 1";
+      return "$fieldValue == 1";
 
       // double, int, String
     } else if (checker.isDartCoreType) {
-      return "data['$name'] as ${field.type}$defaultValue";
+      return "$fieldValue as ${field.type}$defaultValue";
 
       // Iterable
     } else if (checker.isIterable) {
       final argTypeChecker = OfflineFirstChecker(checker.argType);
       final argType = checker.unFuturedArgType;
-      final castIterable = iterableCast(argType,
+      final castIterable = SerdesGenerator.iterableCast(argType,
           isSet: checker.isSet,
           isList: checker.isList,
           isFuture: checker.isArgTypeASibling || checker.isArgTypeAFuture);
@@ -67,7 +64,7 @@ class SqliteDeserialize extends OfflineFirstSerdesGenerator<Sqlite> {
           Query.where('${InsertTable.PRIMARY_KEY_FIELD}', ${InsertTable.PRIMARY_KEY_FIELD}, limit1: true),
         ''';
         final method = '''
-          jsonDecode(data['$name'] ?? []).map((${InsertTable.PRIMARY_KEY_FIELD}) $awaited repository?.getAssociation<$argType>(
+          jsonDecode($fieldValue ?? []).map((${InsertTable.PRIMARY_KEY_FIELD}) $awaited repository?.getAssociation<$argType>(
               $query
             )?.then((r) => (r?.isEmpty ?? true) ? null : r.first)
           )$castIterable
@@ -98,7 +95,7 @@ class SqliteDeserialize extends OfflineFirstSerdesGenerator<Sqlite> {
         if (_hasConstructor) {
           final serializableType = argTypeChecker.superClassTypeArgs.last.getDisplayString();
           return '''
-            jsonDecode(data['$name']).map(
+            jsonDecode($fieldValue).map(
               (c) => $argType.$constructorName(c as $serializableType)
             )$castIterable
           ''';
@@ -107,53 +104,53 @@ class SqliteDeserialize extends OfflineFirstSerdesGenerator<Sqlite> {
 
       // Iterable<DateTime>
       if (argTypeChecker.isDateTime) {
-        return "jsonDecode(data['$name']).map((d) => DateTime.tryParse(d ?? ''))$castIterable";
+        return "jsonDecode($fieldValue).map((d) => DateTime.tryParse(d ?? ''))$castIterable";
       }
 
       // Iterable<enum>
       if (argTypeChecker.isEnum) {
-        return "jsonDecode(data['$name']).map((d) => d as int > -1 ? $argType.values[d as int] : null)$castIterable";
+        return "jsonDecode($fieldValue).map((d) => d as int > -1 ? $argType.values[d as int] : null)$castIterable";
       }
 
       // Iterable<bool>
       if (argTypeChecker.isBool) {
-        return "jsonDecode(data['$name']).map((d) => d == 1)$castIterable";
+        return "jsonDecode($fieldValue).map((d) => d == 1)$castIterable";
       }
 
       // Iterable<double>, Iterable<int>, Iterable<num>, Iterable<Map>, Iterable<String>
-      return "jsonDecode(data['$name'])$castIterable";
+      return "jsonDecode($fieldValue)$castIterable";
 
       // OfflineFirstModel, Future<OfflineFirstModel>
     } else if (checker.isSibling) {
       final query = '''
-        Query.where('${InsertTable.PRIMARY_KEY_FIELD}', data['$name'] as int, limit1: true),
+        Query.where('${InsertTable.PRIMARY_KEY_FIELD}', $fieldValue as int, limit1: true),
       ''';
 
       if (wrappedInFuture) {
-        return '''(data['$name'] > -1
+        return '''($fieldValue > -1
             ? repository?.getAssociation<${checker.unFuturedType}>($query)?.then((r) => (r?.isEmpty ?? true) ? null : r.first)
             : null)''';
       }
 
-      return '''(data['$name'] > -1
+      return '''($fieldValue > -1
             ? (await repository?.getAssociation<${checker.unFuturedType}>($query))?.first
             : null)''';
 
       // enum
     } else if (checker.isEnum) {
-      return "(data['$name'] > -1 ? ${field.type}.values[data['$name'] as int] : null)$defaultValue";
+      return "($fieldValue > -1 ? ${field.type}.values[$fieldValue as int] : null)$defaultValue";
 
       // serializable non-adapter OfflineFirstModel, OfflineFirstSerdes
     } else if (checker.hasSerdes) {
       final _hasConstructor = hasConstructor(field.type);
       if (_hasConstructor) {
         final serializableType = checker.superClassTypeArgs.last.getDisplayString();
-        return "${field.type}.$constructorName(data['$name'] as $serializableType)";
+        return "${field.type}.$constructorName($fieldValue as $serializableType)";
       }
 
       // Map
     } else if (checker.isMap) {
-      return "jsonDecode(data['$name'])";
+      return "jsonDecode($fieldValue)";
     }
 
     return null;