+++ packages/brick_offline_first_abstract/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Remove placeholders from `OfflineFirst` annotation (moves to `FieldSerializable`)
+
 ## 0.0.3
 
 * Rename `ConnectOfflineFirst` as `ConnectOfflineFirstWithRest` to reflect the narrow scope of the annotation